[url_launcher] Use the common PlatformInterface functionality (#2399)
Replace the current mechanism for protecting against implements with the common PlatformInterface class from the plugin_platform_interface package.

Update the unit test to work with the newly used PlatformInterface.
To keep CI green this includes a path dependency override.

This change is going to land and publish with the following sequence:

1. land this PR
2. publish url_launcher_platform_interface
3. land a PR to remove the dependency override from url_launcher, bump it's dependency on url_launcher_platform_interface, and re-enable the publish check for url_launcher
4. publish url_launcher