[google_maps_flutter] Introduce google_maps_flutter_platform_interface. (#2637)
This package contains the platform interface (and the MethodChannel default implementation) for the Google Maps Flutter plugin, up to version 0.5.26.

The most drastic change is including Streams of MapEvents for all different types of events that may come from the native side.

Core types of the plugin (LatLng, Marker, Circle...) now live in this package.

Co-authored-by: chung2012 <chungyl2012@gmail.com>