[in_app_purchase] Rename iOS implementation to _storekit (#4518)
Renames the iOS implementation package to `in_app_purchase_storekit`, and renames it types and files accordingly, to allow for adding a symlink-based macOS implementation to this package in the future.

Temporarily marks `in_app_purchase` as unpublishable so that this can be a move, rather than a copy-and-delete-later, which makes for difficult reviews and destroys git blame.

Part of https://github.com/flutter/flutter/issues/86076
Supports https://github.com/flutter/flutter/issues/84391